requirements.txt
albumentations>=1.3.0
auto_mix_prep>=0.2.0
fastai>=2.7.9
hydra-core>=1.2.0
kornia>=0.6.7
matplotlib>=3.3.4
numpy>=1.20.1
omegaconf>=2.2.1
opencv_python>=4.5.4.60
packaging>=21.0
Pillow>=9.3.0
pytest>=6.2.4
python-dotenv>=0.21.0
pytorch_lightning>=1.6.5
rich>=12.6.0
scikit_learn>=1.1.3
seaborn>=0.11.1
setuptools>=59.5.0
sh>=1.14.2
torch>=1.8.1
torchmetrics>=0.10.2
torchvision>=0.9.1
wandb>=0.10.32
